==> Pandas-Tutorial/0. DataFrame and Series.ipynb <==
## Dictionary

person = {
    'first_name': 'Ali',
    'last_name': 'Bumi',
    'age': 18
}

print(person)

type(person)

people = {
    'first_name': ['Ali', 'Raib', 'Seli'],
    'last_name': ['Bumi', 'Bulan', 'Matahari'],
    'age': [18, 18, 18]
}

people

type(people)

## Data Frame (rows x columns)

import pandas as pd

df = pd.DataFrame(people)

df

type(df)

## Series (rows x one column)

df['first_name']

type(df['first_name'])

df.first_name

df.count

## Multiple Series -> Data Frame

df[['first_name', 'last_name']]

type(df[['first_name', 'last_name']])

## Rows and Columns

df.columns

### iloc (work with index)

#access one row
df.iloc[0]

#acces one row with one column
df.iloc[0]['age']

#access one row with multiple columns
df.iloc[0][['first_name', 'last_name']]

#access one row and one column spesificly using index
print(df.iloc[0, 0])
print(df.iloc[0, 1])
print(df.iloc[0, 2])

#acces multiple rows
df.iloc[[0, 2]]

#access range of rows
df.iloc[0:2]

#access multiple row and one column 
df.iloc[[0, 2], 0]

#access multiple row and multiple column by index
df.iloc[[0, 2], [0, 1]]

#acess multiple row and multiple column by label
df.iloc[[0, 2]][['first_name', 'last_name']]

#acess multiple row by range of index and multiple column by label
df.iloc[0:2][['first_name', 'last_name']]

### loc (work with label)

df.loc[0]

df.loc[[0, 1]]

df.loc[[0, 1], 'age']

df.loc[[0, 1], ['first_name', 'age']]
