==> Pandas-Tutorial/1. Indexes.ipynb <==
import pandas as pd
pd.__version__

people = {
    'first_name': ['Ali', 'Raib', 'Seli'],
    'last_name': ['Bumi', 'Bulan', 'Matahari'],
    'email': ['alibumi@gmail.com', 'raibbintang@gmail.com', 'selimatahari@gmail.com']
}

df = pd.DataFrame(people)

df

df['email']

df.set_index('email')

df

df.set_index('email', inplace=True)

df

df.index

df.loc['alibumi@gmail.com', 'first_name']

df.iloc[0]

df.iloc[0, 0]

df.reset_index(inplace=True)

df
