==> Pandas-Tutorial/2. Filterting.ipynb <==
import pandas as pd
pd.__version__

people = {
    'first_name': ['Ali', 'Raib', 'Seli', 'Ily'],
    'last_name': ['Bumi', 'Bulan', 'Matahari', 'Bulan'],
    'email': ['alibumi@gmail.com', 'raibbintang@gmail.com', 'selimatahari@gmail.com', 'ilybulan@gmail.com']
}

df = pd.DataFrame(people)

df

filt1 = df['last_name'] == 'Bulan'

df[filt1]

df[df['last_name'] == 'Bumi']

df.loc[filt1]

df.loc[filt1, 'email']

filt2 = (df['first_name'] == 'Raib') & (df['last_name'] == 'Bulan')

df.loc[filt2]

filt3 = (df['last_name'] == 'Bumi') | (df['last_name'] == 'Bulan')
df.loc[filt3]

filt4 = df['last_name'] == 'Bulan'
df.loc[-filt4]
