==> Pandas-Tutorial/3. Updating.ipynb <==
import pandas as pd
pd.__version__

people = {
    'first_name': ['Ali', 'Raib', 'Seli', 'Ily'],
    'last_name': ['Bumi', 'Bulan', 'Matahari', 'Bulan'],
    'email': ['alibumi@gmail.com', 'raibbintang@gmail.com', 'selimatahari@gmail.com', 'ilybulan@gmail.com']
}

df = pd.DataFrame(people)

df

## Updating Columns

df.columns

df.columns = ['first_name', 'clan', 'email']

df

df.columns = [x.lower() for x in df.columns]
df

df.columns = df.columns.str.replace(' ', '_')
df

df.rename(columns={'first_name': 'name'}, inplace=True)
df

## Updating Data

### Updating data in multiple columns

df.loc[0] = ['Ali', 'Sagaras', 'alibumi@gmail.com']
df

### Updating data in a single column

df.loc[0, 'email'] = 'alisagaras@gmail.com'
df

df.at[3, 'clan'] = 'Matahari Minor'
df

filt = df['email'] == 'alisagaras@gmail.com'
df[filt]['email'] = 'ali@sagaras.com' #it doesnt work

df.loc[filt, 'email'] = 'ali@sagaras.com' #it works
df

df['email'] = df['email'].str.lower()

df

### Apply - Works for Dataframe and Series

### Apply with series

def update_email(email):
    return email.upper()

#apply -> works for df and series
df['email'] = df['email'].apply(update_email)
df

df['email'] = df['email'].apply(lambda x: x.lower())
df

def change_email(name):
    new_email = name + '@' + 'clan' + '.com'
    return new_email.lower()

df['email'] = df['name'].apply(change_email)
df

df['email'].apply(len)

### Apply with dataframes

#df len
df.apply(len)

df.apply(len, axis='columns')

len(df['email'])

df

df.apply(pd.Series.min)

df.apply(lambda x: x.min())

### ApplyMap -> Apply to each individuals on dataframe 
(only works on dataframes)

df.applymap(len)

df.applymap(str.lower)

### Map -> Apply to each values on series
(only works on series)

df

df['name'].map({'Ali': 'Ali Ceros', 'Raib': 'Putri Raib'})

### Replace -> Substitute only spesific values on series

df['name'].replace({'Ali': 'Ali Ceros', 'Raib': 'Putri Raib'})

df

df.to_csv('people.csv', index=False)
