==> Pandas-Tutorial/4. Adding-Removing.ipynb <==
import pandas as pd
pd.__version__

df = pd.read_csv('data/people.csv')
df

## Adding a column

(df['name'] + '@' + df['clan'] + '.com').str.lower().str.replace(' ', '')

new_email = (df['name'] + '@' + df['clan'] + '.com').str.lower().str.replace(' ', '')

df['new_email'] = new_email
df

## Removing a column

df.drop(columns='email', inplace=True)

df

## Adding multiple columns

name_clan = df['new_email'].str.replace('.com', '').str.split('@', expand=True)

df[['new_name', 'new_clan']] = name_clan
df

## Removing multiple columns

df.drop(columns=['new_name', 'new_clan'], inplace=True)

df

df.rename(columns={'new_email':'email'}, inplace=True)
df

## Adding a row

df = df.append({'name': 'Av'}, ignore_index=True)

df

## Adding multiple rows

people = {
    'name': ['Lumpu', 'Kosong'],
    'clan': ['Nebula', 'Nebula']
}

df2 = pd.DataFrame(people)

df2

df2['email'] = (df2['name'] + '@' + df2['clan'] + '.com').str.lower()

df2

df.append(df2, ignore_index=True)

df3 = pd.concat([df, df2], ignore_index=True)

df3

## Removing a  row

df3.drop(index=4, inplace=True)

df3

filt = (df['clan'] == 'Matahari Minor')

df3.drop(index=df[filt].index)

df3.to_csv('people2.csv', index=False)
