==> Pandas-Tutorial/5. Sorting.ipynb <==
import pandas as pd

df =pd.read_csv('data/people2.csv')
df

## Sorting by a column

df.sort_values(by='name') 

df.sort_values(by='name', ascending=False) 

## Sorting by multiple columns

df.sort_values(by=['name', 'clan'], ascending=False) 

df.sort_values(by=['name', 'clan'], ascending=[True, True], inplace=True) 

df

## Sorting by index

df.sort_index()

## Sorting a series

df['clan'].sort_values()
