==> Pandas-Tutorial/7. Cleaning.ipynb <==
import pandas as pd
import numpy as np

df = pd.read_csv('data/people2.csv')
df

people = {
    'name': ['Nou', 'Tamuz', np.nan],
    'clan': ['Polaris', np.nan, None],
    'email': ['None', 'tamus@bulan.com', 'Missing'],
    'age': [15, 'NA', None]
}

df2 = pd.DataFrame(people)
df2

df3 = pd.concat([df, df2], ignore_index=True)
df3

df3['age'] = [17, 16, 17, 20, 35, 30, 15, 100, None]

df = df3
df

## Droping Missing Values

### Dropping by index

df.dropna()

df.dropna(axis='index', how='any') #default

df.dropna(axis='index', how='all') #drop only if all columns in a row is a missing value

### Dropping by columns

df.dropna(axis='columns', how='any')

df.dropna(axis='columns', how='all')

### Dropping spesific values

df.dropna(axis='index', how='any', subset=['clan'])

df.dropna(axis='index', how='any', subset=['name', 'clan'])

df.dropna(axis='index', how='all', subset=['name', 'clan'])

## Replace missing values

people = {
    'name': ['Nou', 'Tamuz', np.nan],
    'clan': ['Polaris', np.nan, None],
    'email': ['None', 'tamus@bulan.com', 'Missing'],
    'age': [15, 'NA', None]
}

df4 = pd.DataFrame(people)
df4

df5 = df4.replace('NA', np.nan)
df5.replace('Missing', np.nan, inplace=True)
df5

df = pd.concat([df3, df5], ignore_index=True)
df

df.dropna(axis='index', how='all', subset=['name', 'clan'])

## Handling Missing Values

df.isna() #check valus whether is it missing or not?

df.fillna('MISSING') #fill the missing values with a spesifiv value

df['age'].fillna(0)

print(df['age'].mean())
print(df['age'].median())

age_ser = round(df['age'].fillna(df['age'].mean()))
age_ser

df['age'] = age_ser
df

## Casting Datatype Series

df.dtypes

type(np.nan)

df['age'] = df['age'].astype(int)
df

## Dropping Duplicate Rows

df.replace('Missing', np.nan, inplace=True)

# keep first duplicate row
df.drop_duplicates(inplace=True)
df

## Full Cleaning

df.dropna(axis='index', how='any', subset=['name', 'clan'], inplace=True)
df

update_email = (df['name'] + '@' + df['clan'] + '.com').str.replace(' ', '').str.lower()
update_email

df['email'] = update_email
df

df.to_csv('data/people3.csv', index=False)
